requirements.txt
pillow
scikit-learn
matplotlib
numpy
seaborn
opencv-python
tensorflow
